requirements.txt
joblib==1.4.2
kagglehub==0.3.10
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
PyYAML==6.0.2
scikit-learn==1.6.1
streamlit==1.44.1
pytest==8.3.5

